backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
pdf2image==1.17.0
pytesseract==0.3.13
paddleocr
paddlepaddle
Pillow
numpy
